requirements.txt
blinker==1.6.3
click==8.1.7
Flask
gunicorn
importlib-metadata
itsdangerous==2.1.2
Jinja2
MarkupSafe
packaging==23.2
python-dotenv
Werkzeug
zipp
